refactor: Planned/LocalCompanionCard 컴포넌트를 분리하여 구조 개선
