+++ src/features/local-companion/index.js
@@ -1,3 +1,5 @@
 export { default as LocalCompanionPage } from "./pages/LocalCompanionPage";
 
+export { default as LocalCompanionCard } from "./components/LocalCompanionCard";
+
 export { default as CreateLocalModal } from "./modals/LocalCompanionPostModal";
