+++ src/features/local-companion/pages/LocalCompanionPage.jsx
@@ -2,7 +2,10 @@
 
 import { useState } from "react";
 import { Filterbar } from "@/shared";
-import { CreateLocalModal } from "@/features/local-companion";
+import {
+  LocalCompanionCard,
+  CreateLocalModal,
+} from "@/features/local-companion";
 
 const LocalCompanionPage = ({ isLoggedIn }) => {
   const [filters, setFilters] = useState({
@@ -82,46 +85,6 @@ const LocalCompanionPage = ({ isLoggedIn }) => {
           onPostCreate={handlePostCreate}
         />
       )}
-    </div>
-  );
-};
-
-const LocalCompanionCard = ({ title, rating, author, profileImage, tags }) => {
-  const renderStars = (rating) => {
-    return Array.from({ length: 5 }, (_, index) => (
-      <span key={index} className={`star ${index < rating ? "filled" : ""}`}>
-        ★
-      </span>
-    ));
-  };
-
-  return (
-    <div className="local-companion-card">
-      <div className="card-rating">{renderStars(rating)}</div>
-
-      <div className="card-main-content">
-        <h3 className="card-title">{title}</h3>
-
-        <div className="card-author-info">
-          <img
-            src={profileImage || "/images/default-user-profile.png"}
-            alt={author}
-            className="profile-image"
-            onError={(e) => {
-              e.currentTarget.src = "/images/default-user-profile.png";
-            }}
-          />
-          <span className="author-name">{author}</span>
-        </div>
-      </div>
-
-      <div className="card-tags">
-        {tags.map((tag, index) => (
-          <span key={index} className="tag">
-            {tag} ×
-          </span>
-        ))}
-      </div>
     </div>
   );
 };
