+++ src/features/planned-companion/index.js
@@ -1,4 +1,5 @@
-export { default as PlannedCompanionPage } from "./pages/PlannedCompanionPage";
+export { default as PlannedCompanionPage } from "/src/features/planned-companion/pages/PlannedCompanionPage";
 
+export { default as PlannedCompanionCard } from "/src/features/planned-companion/components/PlannedCompanionCard";
 export { default as PlannedCompanionPostModal } from "./modals/PlannedCompanionPostModal";
 export { default as CreatePlannedModal } from "./modals/PlannedCompanionPostModal";
